migrate queue to trpc
