+++ src/server/routers/index.ts
@@ -1,12 +1,16 @@
 import { customAlphabet } from 'nanoid';
 import { z } from 'zod';
 
-import { publicProcedure, router } from '../trpc';
 import supabase from '@/util/supabase';
+import Song from 'src/models/Song';
+
+import { publicProcedure, router } from '../trpc';
+
 const nanoid = customAlphabet(
   '0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz',
   10
 );
+
 const appRouter = router({
   createRoom: publicProcedure
     .input(
@@ -16,11 +20,10 @@ const appRouter = router({
         owner_id: z.string(),
       })
     )
-    .mutation(async (req) => {
+    .mutation(async ({ input }) => {
       const room = {
-        // ...req.input,
-        name: req.input.name,
-        isPublic: req.input.isPublic,
+        name: input.name,
+        isPublic: input.isPublic,
         slug: nanoid(),
       };
 
@@ -28,6 +31,47 @@ const appRouter = router({
 
       if (!res.error) return room;
       return null;
+    }),
+  queueSong: publicProcedure
+    .input(
+      z.object({
+        spotifyUri: z.string().optional(),
+        youtubeVideoID: z.string().optional(),
+        roomId: z.number().optional(),
+        progress: z.number().optional(),
+        duration_ms: z.number().optional(),
+      })
+    )
+    .mutation(async ({ input }) => {
+      const { roomId, spotifyUri, youtubeVideoID, progress, duration_ms } =
+        input;
+
+      if (roomId && (spotifyUri || youtubeVideoID) && duration_ms) {
+        const roomSongResponse = await supabase
+          .from('songs')
+          .select('id')
+          .eq('room_id', roomId);
+        const hasCurrentSong = !!roomSongResponse.data?.length;
+
+        const song: Partial<Song> = {
+          spotifyUri: spotifyUri || undefined,
+          youtube_video_id: youtubeVideoID || undefined,
+          progress: progress || 0,
+          isPaused: hasCurrentSong,
+          room_id: roomId,
+          duration_ms: duration_ms,
+        };
+
+        const songResponse = await supabase.from('songs').insert([song]);
+        if (songResponse.data?.length) {
+          const supabaseSong = songResponse.data[0] as unknown as Song;
+
+          return {
+            room_id: roomId,
+            song_id: supabaseSong.id,
+          };
+        }
+      }
     }),
 });
 export type AppRouter = typeof appRouter;
